pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autowriter"
version = "0.1.0"
description = "Personal novel writing helper with simple project management and AI-friendly hooks."
authors = [{ name = "AutoWriter", email = "author@example.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.23.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "redis>=5.0.0",
    "pydantic-settings>=2.0.0",
    "streamlit>=1.30.0",
    "openai>=1.0.0",
    "google-generativeai>=0.7.0",
]

[project.scripts]
autowriter = "autowriter.cli:main"

[tool.setuptools.packages.find]
where = ["src", "novel_system"]


[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
